requirements.txt
torch==2.6.0
torchmetrics==1.7.0
h5py==3.13.0
lightning==2.5.1
ruamel.yaml==0.18.10 
tensorboard==2.19.0 
